Started comments, added the code of duplicates cable reels

+++ AVCAD Cable Tools/Commands/CableReels/DeleteCableReelCommand.cs
@@ -1,6 +1,7 @@
 ﻿using AVCAD.Models;
 using AVCAD.ViewModels;
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -9,25 +10,46 @@ using System.Windows;
 
 namespace AVCAD.Commands.CableReels
 {
+    /// <summary>
+    /// Command to delete selected cable reel
+    /// </summary>
     public class DeleteCableReelCommand : CommandBase
     {
         private CableReelsPageViewModel cableReelsPageViewModel;
-        private CableReel cableReel;
 
+        /// <summary>
+        /// constructor
+        /// </summary>
+        /// <param name="cableReelsPageViewModel">View model for the page</param>
         public DeleteCableReelCommand(CableReelsPageViewModel cableReelsPageViewModel)
         {
             this.cableReelsPageViewModel = cableReelsPageViewModel;
         }
 
+        /// <summary>
+        /// Execute method.
+        /// </summary>
+        /// <param name="parameter">Collection of CableReelViewModel</param>
         public override void Execute(object? parameter)
         {
+            IEnumerable enumerable = parameter as IEnumerable;
+            if (enumerable == null)
+                throw new ArgumentException("parameter has to be an IEnumerable.", "parameter");
+            //Get selected CableViewModels from the parameter.
+            var selectedReels = enumerable.OfType<ViewModels.CableReelViewModel>().ToList();
+
+            //Ask if user reaaly want to delete this item.
             var dialog = MessageBox.Show("Do you want to delete selected cable reels from the database?", "Warning", MessageBoxButton.YesNo);
             if (dialog == MessageBoxResult.Yes)
             {
-                var selectedCableReelsId = cableReelsPageViewModel.CableReels.Where(i => i.IsSelected).Select(i => i.Id);
+                // Get Ids of the selected elements
+                var selectedCableReelsId = selectedReels.Select(i => i.Id);
+
                 using (var db = new SQlite.ApplicationContext())
                 {
+                    //Find the cableReels in the database
                     var cableReels = db.CableReels.Where(i => selectedCableReelsId.Contains(i.Id));
+                    //Remove, save and update data.
                     foreach (var cableReel in cableReels)
                     {
                         db.CableReels.Remove(cableReel);
